==> notebooks/02_feature_engineering.ipynb <==
# autoreload
%load_ext autoreload
%autoreload 2

# change current working directory to the root of the project
import os
os.chdir(os.path.dirname(os.getcwd()))

# Purpose
- Purpose of this notebook is to finalize features to be used for training

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.feature_selection import mutual_info_regression

import gc
import warnings
from IPython.display import display

# Load Data

nrows = None
# import csv data
df_data = pd.read_csv('data/raw/train_data.csv', nrows=nrows)
df_labels = pd.read_csv('data/raw/train_labels.csv', nrows=nrows)

display(df_data.head())
display(df_labels.head())

# Feature Engineering

def feature_engineering(df: pd.DataFrame, df_labels: pd.DataFrame):
    warnings.filterwarnings("ignore")
    '''Feature Engineering'''

    features = df.drop(["customer_ID", "S_2"], axis = 1).columns.to_list()

    cat_cols = [
        "B_30",
        "B_38",
        "D_114",
        "D_116",
        "D_117",
        "D_120",
        "D_126",
        "D_63",
        "D_64",
        "D_66",
        "D_68"]
    
    #select all columns except cat_cols
    num_cols = [col for col in features if col not in cat_cols]
    
    # feature engineering on categorical columns
    print('Feature Engineering on categorical columns...')
    df_cat_aggregated = df.groupby('customer_ID')[cat_cols].agg(['count', 'last', 'first', 'nunique'])
    df_cat_aggregated.columns = ["_".join(x) for x in df_cat_aggregated.columns]
    df_cat_aggregated.reset_index(inplace = True)

    # feature engineering on numerical columns
    print('Feature Engineering on numerical columns...')
    df_num_aggregated = df.groupby('customer_ID')[num_cols].agg(['last', 'first', 'mean', 'std', 'min', 'max'])
    df_num_aggregated.columns = ["_".join(x) for x in df_num_aggregated.columns]
    df_num_aggregated.reset_index(inplace = True)

    # feature engineering on lag features
    print('Feature Engineering on lag features...')
    for col in df_num_aggregated.columns:
        # subtract columns that have last and first in their name
        if 'last' in col and col.replace('last', 'first') in df_num_aggregated.columns:
            df_num_aggregated[col.replace('last', 'sub')] = df_num_aggregated[col] - df_num_aggregated[col.replace('last', 'first')]
        # divide columns that have last and first in their name
        if 'last' in col and col.replace('last', 'first') in df_num_aggregated.columns:
            df_num_aggregated[col.replace('last', 'frac')] = df_num_aggregated[col] / df_num_aggregated[col.replace('last', 'first')]

    # replace all inf values with 0
    df_num_aggregated.replace([np.inf, -np.inf], 0, inplace=True)

    # join the two dataframes with df_labels on customer_ID
    df = df_labels.merge(df_cat_aggregated, on = 'customer_ID', how = 'left')
    df = df.merge(df_num_aggregated, on = 'customer_ID', how = 'left')

    print("Dimensions after feature engineering", df.shape )

    return df   

# free memory
gc.collect()

df = feature_engineering(df_data, df_labels)

# check length of df == length of df_labels
assert len(df) == len(df_labels)

# Save df

# save df to data/transformed
df.to_csv('data/transformed/df_transformed.csv', index = False)
