==> notebooks/03_train_validation_test_split.ipynb <==
# autoreload
%load_ext autoreload
%autoreload 2

# change current working directory to the root of the project
import os
os.chdir(os.path.dirname(os.getcwd()))

# Purpose
- Purpose of this notebook is to create train, validation and test sets
    - 64% Train Set
    - 8% Validation Set for Calibration
    - 8% Validation Set for Optimal Threshold
    - 20% Test Set

from IPython.display import display

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split

# Load Data

nrows = None
df_data = pd.read_csv("data/transformed/df_transformed.csv", nrows=nrows)
df_data.head()

# Split the data into train, validation and test sets

# Split data into train and test sets
X_train, X_test, y_train, y_test = train_test_split(df_data.drop("target", axis=1), df_data["target"], test_size=0.2, random_state=42)

# Split train set into train and calibration validation sets
X_train, X_val_calibration, y_train, y_val_calibration = train_test_split(X_train, y_train, test_size=0.1, random_state=42)


# Split test set into test and threshold validation sets
X_train, X_val_threshold, y_train, y_val_threshold = train_test_split(X_train, y_train, test_size=0.1, random_state=42)

print('Training Set')
print(f"X_train shape: {X_train.shape}")
print(f"y_train shape: {y_train.shape}")
display(X_train.head())
display(y_train.head())

print('Calibration Validation Set')
print(f"X_val_calibration shape: {X_val_calibration.shape}")
print(f"y_val_calibration shape: {y_val_calibration.shape}")
display(X_val_calibration.head())
display(y_val_calibration.head())

print('Threshold Validation Set')
print(f"X_val_threshold shape: {X_val_threshold.shape}")
print(f"y_val_threshold shape: {y_val_threshold.shape}")
display(X_val_threshold.head())
display(y_val_threshold.head())

print('Test Set')
print(f"X_test shape: {X_test.shape}")
print(f"y_test shape: {y_test.shape}")
display(X_test.head())
display(y_test.head())

# Save data
X_train.to_csv("data/transformed/X_train.csv", index=False)
X_val_calibration.to_csv("data/transformed/X_val_calibration.csv", index=False)
X_val_threshold.to_csv("data/transformed/X_val_threshold.csv", index=False)
X_test.to_csv("data/transformed/X_test.csv", index=False)

y_train.to_csv("data/transformed/y_train.csv", index=False)
y_val_calibration.to_csv("data/transformed/y_val_calibration.csv", index=False)
y_val_threshold.to_csv("data/transformed/y_val_threshold.csv", index=False)
y_test.to_csv("data/transformed/y_test.csv", index=False)
