==> notebooks/06_uncalibrated_model_training.ipynb <==
# autoreload
%load_ext autoreload
%autoreload 2

# change current working directory to the root of the project
import os
os.chdir(os.path.dirname(os.getcwd()))

# Purpose
- Purpose of this notebook is to train an uncalibrated probablistic model using LightGBM

import warnings
import gc
from IPython.display import display

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import joblib

import lightgbm as lgb
import optuna

from sklearn.pipeline import make_pipeline
from sklearn.compose import make_column_transformer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.impute import SimpleImputer
from sklearn.model_selection import KFold
from sklearn.metrics import brier_score_loss

from src.pipeline_transformers import ColumnDropperTransformer

# Load Data

nrows = None

X_train = pd.read_csv("data/transformed/X_train.csv", nrows=nrows)
X_test = pd.read_csv("data/transformed/X_test.csv", nrows=nrows)

y_train = pd.read_csv("data/transformed/y_train.csv", nrows=nrows)
y_test = pd.read_csv("data/transformed/y_test.csv", nrows=nrows)

print(f"X_train shape: {X_train.shape}")
display(X_train.head())
print(f"y_train shape: {y_train.shape}")
display(y_train.head())

print(f"X_test shape: {X_test.shape}")
display(X_test.head())
print(f"y_test shape: {y_test.shape}")
display(y_test.head())

# Change dtypes

cat_cols = [
    "B_30",
    "B_38",
    "D_114",
    "D_116",
    "D_117",
    "D_120",
    "D_126",
    "D_63",
    "D_64",
    "D_66",
    "D_68"]

# select all columns that has cat_cols as prefix
cat_cols = [col for col in X_train.columns if col.startswith("B_30") or col.startswith("B_38") or col.startswith("D_114") or col.startswith("D_116") or col.startswith("D_117") or col.startswith("D_120") or col.startswith("D_126") or col.startswith("D_63") or col.startswith("D_64") or col.startswith("D_66") or col.startswith("D_68")]
# drop columns in cat_cols that has nunique or count in the name
cat_cols = [col for col in cat_cols if not col.endswith("nunique") and not col.endswith("count")]
# add customer_ID to cat_cols
cat_cols.append("customer_ID")

num_cols = [col for col in X_train.columns if col not in cat_cols]

print(f"cat_cols: {cat_cols}")
print(f"num_cols: {num_cols}")

# change dtypes
X_train[cat_cols] = X_train[cat_cols].astype("object")
X_test[cat_cols] = X_test[cat_cols].astype("object")

X_train[num_cols] = X_train[num_cols].astype("float")
X_test[num_cols] = X_test[num_cols].astype("float") 

# free up memory
gc.collect()

# Define Pipeline

def custom_pipeline(**params):
    cat_cols_transformations = make_pipeline(
        ColumnDropperTransformer('customer_ID'),
        SimpleImputer(strategy='most_frequent'),
        OneHotEncoder(handle_unknown='ignore')
    )

    num_cols_transformations = make_pipeline(
        SimpleImputer(strategy='median'),
        StandardScaler()
    )

    cols_trans = make_column_transformer(
        ( cat_cols_transformations, cat_cols),
        (num_cols_transformations, num_cols),
        remainder='passthrough'
        )

    pipeline = make_pipeline(
        cols_trans,
        lgb.LGBMClassifier(objective='binary', is_unbalance=True, n_jobs=10)
        )
    
    return pipeline

# Hyper-parameter Tuning

# Define hyper-parameters & objective function
# define objective function
def objective(trial: optuna.trial.Trial) -> float:
    '''Takes in hyperparameters as input, and trains a model that computes the average validation error based on KFold cross validation'''

    # define hyperparameters
    params = {
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "colsample_bytree": trial.suggest_float("colsample_bytree", 0.2, 1.0),
        "subsample": trial.suggest_float("subsample", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100),   
    }

    k_fold = KFold(n_splits=5)
    scores = []
    for train_index, val_index in k_fold.split(X_train):
        # split data
        X_train_, X_val = X_train.iloc[train_index], X_train.iloc[val_index]
        y_train_, y_val = y_train.iloc[train_index], y_train.iloc[val_index]

        # create model
        pipeline = custom_pipeline(**params)

        # fit model
        pipeline.fit(X_train_, y_train_)

        # compute validation error
        y_pred = pipeline.predict_proba(X_val)[:,1]
        brier_score = brier_score_loss(y_val, y_pred)

        scores.append(brier_score)
    
    return np.mean(scores)

warnings.filterwarnings('ignore')

# optuna study
study = optuna.create_study(direction='minimize', study_name='uncalibrated_lgbm')
study.optimize(objective, n_trials=5)

# print best parameters
best_params = study.best_trial.params
print(f'{best_params=}')

# fit best params on full training set
pipeline = custom_pipeline(**best_params)
pipeline.fit(X_train, y_train)

import joblib
# save model using joblib
joblib.dump(pipeline, 'models/uncalibrated_lightgbm.pkl')

# load model using joblib
pipeline = joblib.load('models/uncalibrated_lightgbm.pkl')

# compute test error on test set
uncalibrated_predictions = pipeline.predict_proba(X_test)[:,1]

display(y_test.head())
display(pd.DataFrame(uncalibrated_predictions).head())

test_brier_score = brier_score_loss(y_test, uncalibrated_predictions)
print(f'{test_brier_score=:.4f}')

# Save predictions to csv
df_predictions = pd.DataFrame(uncalibrated_predictions, columns=["predictions"])
df_predictions.to_csv("predictions/predictions_uncalibrated_lightgbm_probabilities.csv", index=False)
